added logos

+++ lib/services/post_model.dart
@@ -1,10 +1,10 @@
 class Post {
   final String id;
-  final int price;
+  final double price;
 
   Post({required this.id, required this.price});
 
   factory Post.fromJson(Map<String, dynamic> json) {
-    return Post(id: json['bitcoin'] as String, price: json['price'] as int);
+    return Post(id: json['bitcoin'] as String, price: json['price'] as double);
   }
 }
